refactor: move the Zustand stores out of assets/
assets/ otherwise holds images, fonts and icons. The stores were presumably
filed there once and PROJECTSTRUCTURE.md updated to match, which is how
"Images, fonts, videos, icons, etc. and Zustand store" came to be a sentence.

Pure path change; the imports were about to change anyway in #183.
