+++ components/settings/ThemeSelector.tsx
@@ -1,6 +1,6 @@
 import { View } from "react-native";
 import { Button, List, Text, useTheme } from "react-native-paper";
-import { useThemeColor } from "@/assets/src/calendar-storage";
+import { useThemeColor } from "@/stores/calendar-storage";
 import React, { useEffect } from "react";
 import { ThemedView } from "@/components/ThemedView";
 import { insertSetting, getSetting } from "@/db/database";
